refactor(gfql): extract fast-path specialization helpers to gfql_fast_paths.py
Pure code move, no behavior change. gfql_unified.py had grown to ~3.7k lines as
the connected-join (single/two-star) and grouped-aggregate / two-hop-count fast
paths accumulated across #1730 and #1731. Move both families (22 functions,
~1.63k lines) into a new sibling module graphistry/compute/gfql_fast_paths.py so
the orchestrator stays readable.

- One-directional import: gfql_unified -> gfql_fast_paths (imports the 4 entry
  points it dispatches to); the module imports only leaf modules
  (Engine/ast/chain/filter_by_dict/dataframe/lowering) — no back-edge, no cycle.
- Within the module, the aggregate family calls the connected-join family; both
  self-contained (verified none reference code that stayed in gfql_unified).
- `# pragma: no cover` markers on polars/cuDF-only helpers move with them.
- test_lowering.py import repointed to gfql_fast_paths (only test referencing the
  moved symbols).

Validated on dgx (26.02-gfql-polars): import-clean (no circular/missing);
test_lowering.py 1383 passed/0 failed; polars cypher conformance + matrix 406
passed/0 failed. ruff clean; mypy 0-new (the 11 errors are pre-existing in
untouched polars/row_pipeline.py + degrees.py).
